% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_refit.R
\name{find_refit_fn}
\alias{find_refit_fn}
\title{Find a working refitting function
Try to refit the model using the previously used response}
\usage{
find_refit_fn(object, y)
}
\description{
Find a working refitting function
Try to refit the model using the previously used response
}
\keyword{internal}
