% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coefs_and_vcov.R
\name{get_fixef}
\alias{get_fixef}
\alias{get_fixef.default}
\alias{get_fixef.merMod}
\alias{get_fixef.glmmTMB}
\title{Get fixed effects}
\usage{
get_fixef(object)

\method{get_fixef}{default}(object)

\method{get_fixef}{merMod}(object)

\method{get_fixef}{glmmTMB}(object)
}
\arguments{
\item{object}{A model object for which fixed effects are to be retrieved.}
}
\value{
A numeric vector of fixed effects coefficients.
}
\description{
Extracts the fixed effects coefficients from a model.
}
\details{
By default it calls \code{\link[stats:coef]{stats::coef()}}. If the model class is \code{merMod} calls
\code{fixef}.
}
