% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_response.R
\name{get_model_response}
\alias{get_model_response}
\title{Get Model Response}
\usage{
get_model_response(object)
}
\arguments{
\item{object}{Model to extract responses from.}
}
\description{
Get Model Response
}
