% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_refit.R
\name{get_refit}
\alias{get_refit}
\title{Refit the Model}
\usage{
get_refit(object, newresp, ...)
}
\arguments{
\item{object}{A model.}

\item{newresp}{the new response, may be a vector or a matrix.}

\item{...}{other arguments passed to \code{refit} or \code{update}.}
}
\value{
An updated model object, refitted using the new response variable.
}
\description{
Refit an existing regression model using a new response variable.
}
\details{
This generic method refit the \code{object} with the newresp.
It shold maintain the \code{object} properties by default, but some aspects of
the default fitting method may be overwritten with the \code{...}.
}
\seealso{
\code{\link[stats:update]{stats::update()}}, \code{\link[lme4:refit]{lme4::refit()}}
}
