% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coefs_and_vcov.R
\name{get_vcov}
\alias{get_vcov}
\alias{get_vcov.default}
\alias{get_vcov.glmmTMB}
\title{Get covariance matrix}
\usage{
get_vcov(object)

\method{get_vcov}{default}(object)

\method{get_vcov}{glmmTMB}(object)
}
\arguments{
\item{object}{A model object for which the covariance matrix is to be retrieved.}
}
\value{
A covariance matrix of the model object.
}
\description{
Retrieves the covariance matrix for the fixed effects.
}
\details{
By default it calls \code{\link[stats:vcov]{stats::vcov()}} to retrieve the covariances.
}
