% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_refit.R
\name{refit_safely}
\alias{refit_safely}
\title{Calls f without throwing errors}
\usage{
refit_safely(.f, newresp, ...)
}
\arguments{
\item{.f}{a function that refit a model.}

\item{newresp}{response variable used to fit the model.}

\item{...}{arguments passed to other methods.}
}
\description{
Calls f without throwing errors
}
\keyword{internal}
