% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_refit.R
\name{get_refit}
\alias{get_refit}
\alias{get_refit.merMod}
\alias{get_refit.glmmTMB}
\title{Refit the Model}
\usage{
get_refit(object, new_response, ...)

\method{get_refit}{merMod}(object, new_response, ...)

\method{get_refit}{glmmTMB}(object, new_response, ...)
}
\arguments{
\item{object}{A model.}

\item{new_response}{the new response, may be a vector or a matrix.}

\item{...}{other arguments passed to \code{refit} or \code{update}.}
}
\description{
This function is supposed to be used internally. Use \code{\link[=refit_model]{refit_model()}} instead.
}
\seealso{
\code{\link[stats:update]{stats::update()}}
}
\keyword{internal}
